# By Pascal Obry (9) and others
# Via Ger (1) and others
* upstream/master: (39 commits)
  wb_presets: fix 7D Mark II (#10243)
  Fix reference of wb_presets.c in POFILES.in.
  Add the Pentax *ist DL2 do rawspeed
  snapshots: fix drawing offset for new border system. fixes #10228
  add "// clang-format {off,on}" to nvidia_gpus.h
  Update beautify_style.sh
  Move wb_presets.c to external/
  dt_imageio_export_with_flags(): do not leak memory upon error. Fixes Coverity CID 1260259
  Some more const where possible.
  Diferent black points for samsung NX1 high iso
  Minor code rewrite to avoid distributed complexity.
  Use const qualifier where possible.
  [usermanual] Make sure URLs to external HTML files aren't getting .php appended
  Add some const qualifier where possible.
  Updated Dutch translation
  Update French translation.
  Improve tooltip text for style append.
  Graduatednd IOP: avoid rounding issues for rotation after moving whole line. Fixes #10241
  basecurve: cleanup table layout butched by clang-format
  exif: always try to use Exiv2's lens detection for Olympus
  ...

Conflicts:
	src/imageio/format/exr.cc
	src/imageio/format/j2k.c
	src/imageio/format/jpeg.c
	src/imageio/format/png.c
	src/imageio/format/tiff.c
	src/imageio/format/webp.c
	src/libs/export.c

+++ src/develop/pixelpipe_hb.c
@@ -777,8 +777,9 @@ static int dt_dev_pixelpipe_process_rec(dt_dev_pixelpipe_t *pipe, dt_develop_t *
 #pragma omp parallel for schedule(static) default(none) shared(pipe, roi_out, roi_in, output)
 #endif
           for(int j = 0; j < cp_height; j++)
-            memcpy(((char *)*output) + bpp * j * roi_out->width,
-                   ((char *)pipe->input) + bpp * (in_x + (in_y + j) * pipe->iwidth), bpp * cp_width);
+            memcpy(((char *)*output) + (size_t)bpp * j * roi_out->width,
+                   ((char *)pipe->input) + (size_t)bpp * (in_x + (in_y + j) * pipe->iwidth),
+                   (size_t)bpp * cp_width);
         }
         else
         {
